refactor!: separate code into library and binary crate
This commit separates the code into two crates in a Cargo workspace.
This permits the library crate (called `typescript_tools`) to be more
permissive with its version ranges.

The binary crate is not published to a Cargo registry. The recommended
way to obtain a binary is via the precompiled releases in GitHub Actions
or via the Nix derivation.

BREAKING CHANGE: separate the binary code out of the library crate